requirements.txt
pandas
numpy
nltk
scikit-learn
scipy
matplotlib
seaborn
textblob
textblob-fr
vaderSentiment-fr
sentence-transformers
umap-learn
bertopic
